fix: the final-audit memory-safety family (final audit)
Seven confirmed lifetime/memory fixes, one commit since they share the
ASan+UBSan verification pass:

- fw-cache: the per-page rendering bool is now an in-flight job count
  (plus the render_gen of the newest job). fw_cache_start used to
  blanket-reset the bool to allow re-submission while an old-generation
  worker could still be rendering with that page's parsed handle; the
  Tier-1 eviction trusted the bool and freed the handle mid-use. A
  count only reaches zero when every pushed job has finished, so the
  eviction guard is airtight; re-submission now keys on the job's
  generation instead of a reset.
- fw-window: ExtractCtx owns refs to the window and captures the
  document. The old raw window pointer (whose 'weak: not refed'
  comment manufactured a lifetime guarantee that did not exist)
  dangled if the window closed while the portal folder dialog was up;
  and the window could open a different file mid-dialog, so the
  attachments now explicitly belong to the captured document. A
  summary is only shown if the window is still visible.
- fw-reflow-document-mobi: get_resources returns the bare table per
  the transfer-none contract; the extra ref leaked the resource table
  (and its flow CSS) on every MOBI/AZW3 open.
- fw-reflow-document-epub/-fb2: archive entry claims are capped at the
  CBR backend's 256 MB before g_malloc.
- fw-reflow-document-mobi: inline <style> text containing '</style'
  is refused, as the EPUB collector already did; it would break out of
  the emitted block into the head, where body-level script-stripping
  never runs.
- fw-mobi-parser: parse_indx bounds-checks hdr_length in gsize; the
  32-bit sum wrapped for values >= 0xFFFFFFF4 and fed the TAGX probe
  a wild pointer.
- fw-document-cbr: fz_try-crossing locals (buf/img/cairo_pix/draw_dev/
  surface, the two dim probes' defaults) are volatile, and the file
  header carries the same CRITICAL RULES setjmp block as the PDF
  backend.

Suite: 8/8 native; under ASan+UBSan 7/8 with only the documented
corpus-soak RSS-cap exception (failures=0, peak 2246 vs 1800 MB cap),
matching the recorded sanitizer posture.

+++ src/fw-mobi-parser.c
@@ -639,7 +639,10 @@ parse_indx (const guchar *data, gsize raw_len,
   guint32 num_records = read_be32 (data + off + 24);
   /* num_cncx at 0x34 — not used here (TOC-only). */
 
-  if (hdr_length + 12 > len) return NULL;
+  /* gsize arithmetic: hdr_length is file-controlled and the 32-bit sum
+   * wraps for values ≥ 0xFFFFFFF4, passing this check and feeding the
+   * TAGX probe a wild pointer. */
+  if ((gsize) hdr_length + 12 > len) return NULL;
 
   /* TAGX section starts at indx.length, has its own header
    * (magic 'TAGX', length at 4, num_control_bytes at 8). */
